faker/requirements.txt
faker==19.13.0
psycopg2-binary==2.9.9 